requirements.txt
Flask
tensorflow-cpu
bootstrap-flask